final commit

+++ TP2/TP2/InputHandler.cs
@@ -4,7 +4,7 @@ namespace TP2
 {
     public class InputHandler
     {
-        private const double MONTANT_MIN = 50000;
+        private const decimal MONTANT_MIN = 50000;
         private const int DUREE_MIN = 9 * 12;
         private const int DUREE_MAX = 25 * 12;
 
@@ -15,7 +15,7 @@ namespace TP2
                 throw new ArgumentException("Il faut 3 arguments : montant, durée et taux nominal");
             }
 
-            if (!double.TryParse(args[0], out double montant) || montant < MONTANT_MIN)
+            if (!decimal.TryParse(args[0], out decimal montant) || montant < MONTANT_MIN)
             {
                 throw new ArgumentException("Le montant doit être un nombre supérieur à 50 000 euros");
             }
@@ -25,7 +25,7 @@ namespace TP2
                 throw new ArgumentException("La durée doit être un nombre entier compris entre 9 et 25 ans");
             }
 
-            if (!double.TryParse(args[2], out double tauxNominal))
+            if (!decimal.TryParse(args[2], out decimal tauxNominal))
             {
                 throw new ArgumentException("Le taux nominal doit être un nombre");
             }
